AI-powered-lead-automation-chatbot/requirements.txt

Flask==2.3.3
Flask-Cors==4.0.0
gspread==5.12.0
oauth2client==4.1.3
python-dotenv==1.0.1
requests==2.31.0
phonenumbers==8.13.34
